Make loopback support modular

+++ src/loopback.c
@@ -10,8 +10,20 @@
 #include "shared.h"
 #include "loopback.h"
 
-void loopback_init (loopback_ctx_t *loopback_ctx)
+void loopback_init (loopback_ctx_t *loopback_ctx, const user_options_t *user_options)
 {
+  loopback_ctx->enabled = false;
+
+  if (user_options->benchmark   == true) return;
+  if (user_options->keyspace    == true) return;
+  if (user_options->left        == true) return;
+  if (user_options->show        == true) return;
+  if (user_options->stdout_flag == true) return;
+  if (user_options->usage       == true) return;
+  if (user_options->version     == true) return;
+
+  loopback_ctx->enabled = true;
+
   loopback_ctx->fp = NULL;
 
   loopback_ctx->filename = (char *) mymalloc (HCBUFSIZ_TINY);
@@ -19,18 +31,26 @@ void loopback_init (loopback_ctx_t *loopback_ctx)
 
 void loopback_destroy (loopback_ctx_t *loopback_ctx)
 {
+  if (loopback_ctx->enabled == false) return;
+
   myfree (loopback_ctx->filename);
+
+  myfree (loopback_ctx);
 }
 
-int loopback_write_open (loopback_ctx_t *loopback_ctx, const char *induction_directory)
+int loopback_write_open (loopback_ctx_t *loopback_ctx, const induct_ctx_t *induct_ctx)
 {
+  if (loopback_ctx->enabled == false) return 0;
+
+  if (induct_ctx->enabled == false) return 0;
+
   time_t now;
 
   time (&now);
 
   const uint random_num = get_random_num (0, 9999);
 
-  snprintf (loopback_ctx->filename, HCBUFSIZ_TINY - 1, "%s/%s.%d_%u", induction_directory, LOOPBACK_FILE, (int) now, random_num);
+  snprintf (loopback_ctx->filename, HCBUFSIZ_TINY - 1, "%s/%s.%d_%u", induct_ctx->root_directory, LOOPBACK_FILE, (int) now, random_num);
 
   loopback_ctx->fp = fopen (loopback_ctx->filename, "ab");
 
@@ -46,6 +66,8 @@ int loopback_write_open (loopback_ctx_t *loopback_ctx, const char *induction_dir
 
 void loopback_write_unlink (loopback_ctx_t *loopback_ctx)
 {
+  if (loopback_ctx->enabled == false) return;
+
   if (loopback_ctx->filename == NULL) return;
 
   unlink (loopback_ctx->filename);
@@ -53,6 +75,8 @@ void loopback_write_unlink (loopback_ctx_t *loopback_ctx)
 
 void loopback_write_close (loopback_ctx_t *loopback_ctx)
 {
+  if (loopback_ctx->enabled == false) return;
+
   if (loopback_ctx->fp == NULL) return;
 
   fclose (loopback_ctx->fp);
@@ -60,6 +84,8 @@ void loopback_write_close (loopback_ctx_t *loopback_ctx)
 
 void loopback_format_plain (loopback_ctx_t *loopback_ctx, const u8 *plain_ptr, const unsigned int plain_len)
 {
+  if (loopback_ctx->enabled == false) return;
+
   int needs_hexify = 0;
 
   for (uint i = 0; i < plain_len; i++)
@@ -98,6 +124,8 @@ void loopback_format_plain (loopback_ctx_t *loopback_ctx, const u8 *plain_ptr, c
 
 void loopback_write_append (loopback_ctx_t *loopback_ctx, const u8 *plain_ptr, const unsigned int plain_len)
 {
+  if (loopback_ctx->enabled == false) return;
+
   FILE *fp = loopback_ctx->fp;
 
   loopback_format_plain (loopback_ctx, plain_ptr, plain_len);
